Move utility modules to utils.py
